Fix docker image naming for OSS-Fuzz images (#291)
Correct the runner image used on runner bots to use the benchmark-based naming scheme introduced in #254.

It also removes the oss-fuzz prefix from OSS-Fuzz builder (not including the project builder image that is part of OSS-Fuzz and not fuzzbench) and runners to be consistent with fuzzbench fuzzers.
This fixes #288. and likely fixes #285 (they are probably the same issue).
Also fix some stale names that refer to oss_fuzz_project when they should refer to benchmark.