Add initial support for preemptible trial runners (#302)
This change will use preemptible GCE instances when preemptible_runners: true is specified in experiment-config.yaml.
Using preemptible instances is much cheaper but comes with the following cost:

1. Trials can last only up to 24 hours (probably less if we include start up time)
2. Some trials will be preempted and will not complete.
This patch doesn't change parts of the scheduler that assume that it shuts down all runners. As a result if a trial is preempted the scheduler (and the measurer) will run forever. Thus the dispatcher will need to be terminated manually. For large experiments, the cost of the dispatcher running for some extra time unnecessarily is unlikely to outweigh the savings from using preemptible instances. Proper support
for preemptibles will come in a later CL.