Add src_analysis module (#278)
Add src_analysis module

Add src_analysis module for doing analysis on the fuzzbench source
code. This module can determine which fuzzers and benchmarks have changed.
Use it to build more selectively in CI.

The effect of this that most people notice are:
1. When integrating a benchmark, they will (probably) only see failures related to that benchmark.
2. When integrating a fuzzer they will (probably) only see failures related to that fuzzer.
3. CI builds will start faster in general. That is because the previous system can overload
the limits of CI causing builds to get queued while others finish. This is unlikely to happen anymore.

In general, we won't build fuzzers with all benchmarks in CI when
benchmarks/* or fuzzers/* is changed, unless:
1. A fuzzer has changed. In this case build the fuzzer with all benchmarks, any of them
could be broken.
2. A benchmark has changed . In this case build the benchmark with all fuzzers.
3. Build infra (such as docker/*) has changed. In this case build everything, 
since anything can be broken.

We've made AFL an exception to this. AFL will be built all the time with all benchmarks.
This is supposed to be a backup in case our analysis was incorrect. 
AFL is unlikely to fail spuriously and is pretty fast. But we may remove it if it confuses people
and/or we find it unnecessary and/or it clogs up builds, causing a queue to form.

In the future, we may be able to use this module for determining what should run in experiments.

Also speed up pytest collection.