Retry preemptible trials when preemptible_trials is True (#352)
Retry preemptible trials based on our heuristic that bounds the number of
retries we do and the length of time we do them for to avoid worst-case
scenarios while keeping the average use of preemptibles resource-efficient.

Also:
Add library for using GCE API
Log subcomponent so we can distinguish things logged by scheduler and those
logged by other parts of the dispatcher like the measurer etc.
Don't log an error when generating a report during an experiment if the data
is empty. This is expected before measuring happens.