Set up new queue-based architecture development (#521)
We are implementing a new architecture for FuzzBench, which will be more scalable and completely platform independent. While the [current architecture](https://google.github.io/fuzzbench/reference/how-it-works/) heavily relies on Google Compute Engine (e.g., it directly creates and manages GCE instances), the new architecture is a task-queue based system that can be run on any platform or environment. This includes a local machine, a small lab cluster, or the cloud. Here's a high level overview:
![image](https://user-images.githubusercontent.com/1070168/87357044-374bb800-c531-11ea-8fef-653822f5fc56.png)

The current architecture's main modules (such as the Builder, Dispatcher, Scheduler, and Measurer) will be replaced by the new `run_experiment` and `worker` containers. In the new architecture, deployment (i.e., infrastructure management and container orchestration) will be fully separated from the application logic (e.g., in the local machine use case we will use Docker Compose for orchestration). Our goal is to enable the incremental development of the new architecture's modules, while keeping the old ones working. When the new modules are fully functional, the old ones can be removed. This PR sets up an initial infrastructure for the development of this new architecture, in particular: 
- Add "skeleton" code for the main `run_experiment` module and "fake jobs" code for the `worker`. 
- Add Dockefile for the fuzzbench container image (used by `run_experiment` and `worker`).
- Add docker-compose file to define the local setup for a fuzzbench experiment.
- Add Makefile target for running a local experiment.
- Add redis and rq as dependencies.

To run the experiment skeleton locally, simply run:
```shell
$ make run-experiment
```