Initial prototype of build logic consolidation (#543)
Running a FuzzBench experiment involves building hundreds of docker images for
all the different fuzzer-benchmark pairs. Currently, we build these images using
Google Cloud Build in prod, and using Makefiles locally. The logic of building
the images and the dependencies between them is currently encoded in two places,
ie., in docker/gcb/*.yaml and docker/generate_makefile.py. This is difficult to
maintain and [very error prone](https://github.com/google/fuzzbench/issues/336).
To complicate things further, we are also implementing distributed builds as
part of a new queue-based architecture (https://github.com/google/fuzzbench/pull/521).

To eliminate the above described redundancy and make the FuzzBench build system
simple and easy to maintain, we introduce a `get_images_to_build(fuzzers,
benchmarks)` function that returns the set of buildable images with their
properties describing how to build them, as well as their dependencies (as
python objects). This can be used to simply generate the input for different
build tools, ie. the 1) GCB build spec, 2) the Makefile, and 3) the (rq) jobs to
enqueue in the new architecture.

The PR includes the initial prototype of:
- the `get_images_to_build` function, which relies on
- an easy to read/modify config file describing the
  image types and their dependencies
- generator for Cloud Build spec
- generator for Makefile rules