Fix issue formatting in logs.error. (#663)
Fix issue formatting in logs.error.

There is an issue with our logging module, it doesn't properly handle
messages containing % signs, since it assumes those are formatting
directives (e.g. "%s").

Make it OK to pass unescaped messages to logs.error as long as
there are no actual formatting args by only doing formatting
if formatting args are supplied. We did this previously for other
log messages so this commit does it for error messages.

The particular issue that brought this to light was a git
command from oss_fuzz_benchmark_integration.py. This command
contains a percent sign.