new summary table (#1100)
* new summary table

The new experiment summary table now shows the normalized
relative median performance for fuzzers.  Additionally, the
fuzzers are sorted by their average normalized median coverage,
and the median and mean are added as a summary row.  I think
this is a much more useful summary than the raw median numbers
previously listed.  The top scores are highlighted in green, and
scores greater than 95% (of the experiment max) are highlighted
with blue gradients.

* adding a new summary table for total unique bugs found

- updating summary table(s) for bug reports.
    - Include both code and bug experiment summary table.
    - Adds a total unique bugs found summary table.  This
    information is not found anywhere else in the report.
    Different bugs are found in different trials, but the charts
    don't have a way to report that information.A

* Adds text to summary table to make it more intuitive.

Co-authored-by: Laszlo Szekeres <lszekeres@google.com>
Co-authored-by: Abhishek Arya <inferno@chromium.org>