Start Recording Coverage at Cycle 0 (#1642)
Hello @Alan32Liu and Fuzzbench maintainers!

The goal of this PR is to start recording the coverage statistics at
cycle 0, before the fuzzers are started up. To do so, I'm copying over
the input corpus to the output directory immediately and calling do_sync
before starting the fuzzer thread. Additionally, I adjusted both the
logscale and linear time plots to start at t=0.

Overall, I believe this makes the linear plots quite a bit better, as we
can now see the starting coverage for a given run. There is some
distortion around zero for the logscale plots, happy to discuss how best
to handle this if needed.

These changes also relate to
https://github.com/google/fuzzbench/issues/1489 in that we have observed
the initial coverage to have an impact on the relative performance of
fuzzers in Fuzzbench. Archiving the initial corpus along with the other
snapshots of coverage for the run and recording the initial edges
covered is a first step towards accounting for the impacts of additional
variables on benchmarking.

I've attached some example plots below:

![linear_time](https://user-images.githubusercontent.com/17279039/214587615-863febc6-a684-49f6-b588-a891b9022367.png)

![log_time](https://user-images.githubusercontent.com/17279039/214587626-18d61d63-3f7e-42b3-b37d-7990dde42656.png)

---------

Co-authored-by: van Hauser <vh@thc.org>