afl++ version comparison (#1798)
This run
https://www.fuzzbench.com/reports/experimental/2023-03-23-various/index.html
showed that v3.14c has a better performance than the current 4.06a which
is baffling. The difference is huge!

so this experiment tests all afl++ versions starting with 3.13 up to
current with the same llvm version (it needs version 13 because 3.14 is
obviously not updated to current llvm changes).

It might be the llvm version it might be something else ... no clue, I
need more data.
Curous that this did not come up earlier as I test this rigorously
before a release.